added math test

+++ src/main/java/com/kilowatt/Compiler/Builtins/Libraries/Std/Math/StdMathExt.java
@@ -11,13 +11,13 @@ import com.kilowatt.WattVM.VmAddress;
 public class StdMathExt {
     public static Number cos(Number value) {
         if (value instanceof Integer) {
-            return Math.cos(value.intValue());
+            return (float) Math.cos(value.intValue());
         } else if (value instanceof Long) {
-            return Math.cos(value.longValue());
+            return (float) Math.cos(value.longValue());
         } else if (value instanceof Float) {
             return Math.cos(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.cos(value.doubleValue());
+            return (float) Math.cos(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -31,13 +31,13 @@ public class StdMathExt {
 
     public static Number abs(Number value) {
         if (value instanceof Integer) {
-            return Math.abs(value.intValue());
+            return (float) Math.abs(value.intValue());
         } else if (value instanceof Long) {
-            return Math.abs(value.longValue());
+            return (float) Math.abs(value.longValue());
         } else if (value instanceof Float) {
             return Math.abs(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.abs(value.doubleValue());
+            return (float) Math.abs(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -51,13 +51,13 @@ public class StdMathExt {
 
     public static Number floor(Number value) {
         if (value instanceof Integer) {
-            return value.intValue();
+            return (float) value.intValue();
         } else if (value instanceof Long) {
-            return Math.floor(value.longValue());
+            return (float) Math.floor(value.longValue());
         } else if (value instanceof Float) {
-            return Math.floor(value.floatValue());
+            return (float) Math.floor(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.floor(value.doubleValue());
+            return (float) Math.floor(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -71,11 +71,11 @@ public class StdMathExt {
 
     public static Number sqrt(Number value) {
         if (value instanceof Integer) {
-            return Math.sqrt(value.intValue());
+            return (float) Math.sqrt(value.intValue());
         } else if (value instanceof Long) {
-            return Math.sqrt(value.longValue());
+            return (float) Math.sqrt(value.longValue());
         } else if (value instanceof Float) {
-            return Math.sqrt(value.floatValue());
+            return (float) Math.sqrt(value.floatValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -89,13 +89,13 @@ public class StdMathExt {
 
     public static Number sin(Number value) {
         if (value instanceof Integer) {
-            return Math.sin(value.intValue());
+            return (float) Math.sin(value.intValue());
         } else if (value instanceof Long) {
-            return Math.sin(value.longValue());
+            return (float) Math.sin(value.longValue());
         } else if (value instanceof Float) {
-            return Math.sin(value.floatValue());
+            return (float) Math.sin(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.sin(value.doubleValue());
+            return (float) Math.sin(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -109,13 +109,13 @@ public class StdMathExt {
 
     public static Number tan(Number value) {
         if (value instanceof Integer) {
-            return Math.tan(value.intValue());
+            return (float) Math.tan(value.intValue());
         } else if (value instanceof Long) {
-            return Math.tan(value.longValue());
+            return (float) Math.tan(value.longValue());
         } else if (value instanceof Float) {
-            return Math.tan(value.floatValue());
+            return (float) Math.tan(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.tan(value.doubleValue());
+            return (float) Math.tan(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -129,13 +129,13 @@ public class StdMathExt {
 
     public static Number asin(Number value) {
         if (value instanceof Integer) {
-            return Math.asin(value.intValue());
+            return (float) Math.asin(value.intValue());
         } else if (value instanceof Long) {
-            return Math.asin(value.longValue());
+            return (float) Math.asin(value.longValue());
         } else if (value instanceof Float) {
-            return Math.asin(value.floatValue());
+            return (float) Math.asin(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.asin(value.doubleValue());
+            return (float) Math.asin(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -149,13 +149,13 @@ public class StdMathExt {
 
     public static Number acos(Number value) {
         if (value instanceof Integer) {
-            return Math.acos(value.intValue());
+            return (float) (Math.acos(value.intValue()));
         } else if (value instanceof Long) {
-            return Math.acos(value.longValue());
+            return (float) Math.acos(value.longValue());
         } else if (value instanceof Float) {
-            return Math.acos(value.floatValue());
+            return (float) Math.acos(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.acos(value.doubleValue());
+            return (float) Math.acos(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
@@ -169,13 +169,13 @@ public class StdMathExt {
 
     public static Number atan(Number value) {
         if (value instanceof Integer) {
-            return Math.atan(value.intValue());
+            return (float) Math.atan(value.intValue());
         } else if (value instanceof Long) {
-            return Math.atan(value.longValue());
+            return (float) Math.atan(value.longValue());
         } else if (value instanceof Float) {
-            return Math.atan(value.floatValue());
+            return (float) Math.atan(value.floatValue());
         } else if (value instanceof Double) {
-            return Math.atan(value.doubleValue());
+            return (float) Math.atan(value.doubleValue());
         } else {
             VmAddress address = WattCompiler.vm.getLastCallAddress();
             throw new WattRuntimeError(
